Alterações de conexão correta ao servidor. Pendente ainda de receber corretamente o jobdescription (E aguns outros pacotes) que possuem elementos recentemente implementados e que ainda não foram mapeados em C#.
TODO: Traduzir os types do dofus invoker para C#.

+++ BlueSheep/Common/Protocol/messages/game/character/choice/CharacterFirstSelectionMessage.cs
@@ -32,7 +32,7 @@ namespace BlueSheep.Common.Protocol.Messages
         {
         }
         
-        public CharacterFirstSelectionMessage(int id, bool doTutorial)
+        public CharacterFirstSelectionMessage(ulong id, bool doTutorial)
          : base(id)
         {
             this.doTutorial = doTutorial;
